G: added Qt/Equaliser/CMakeLists.txt G: added Qt/Equaliser/loess.h G: added Qt/Equaliser/main.cpp G: added Qt/Equaliser/mainwindow.cpp G: added Qt/Equaliser/mainwindow.h G: added Qt/Equaliser/mainwindow.ui G: added Qt/Equaliser/network.h G: added Qt/Equaliser/plot.h G: added Qt/Equaliser/runQtCreator.bat G: added Qt/Equaliser/set64bitPath.bat G: added Qt/Equaliser/tiff.cpp G: added bin/am_tiffer.bat G: added bin/am_tiffer_slave.bat G: added bin/equaliser.bat G: added bin/equaliser_slave.bat G: added bin/imageviewer.bat G: added bin/imageviewer_slave.bat G: added loess/CMakeLists.txt G: added loess/loess.cpp G: added loess/loess.h G: added loess/throw.h G: added loess/varkitversion.h G: changed Qt/Compresser/mainwindow.cpp G: changed Qt/Compresser/mainwindow.ui G: changed Qt/Imageviewer/imagewidget.cpp G: changed Qt/Imageviewer/imagewidget.h G: changed Qt/Vdistancer/mainwindow.cpp G: changed Qt/Vdistancer/mainwindow.h G: changed Qt/Vdistancer/mainwindow.ui G: changed bin/info/vdistancer_info.txt G: changed bin/set64bitPath.bat G: changed compress/compress.cpp G: changed connect/connect.cpp G: changed crop/crop.cpp G: changed fill/fill.cpp G: changed smooth/smooth.cpp G: changed threshold/threshold.cpp G: changed topo/topo.cpp G: changed uncompress/CMakeLists.txt G: changed uncompress/uncompress.cpp G: changed vdistancef/vdistance.f90 G: removed equalize/CMakeLists.txt G: removed equalize/equalize.cpp

+++ connect/connect.cpp
@@ -37,7 +37,7 @@ using namespace std;
 
 typedef itk::Image<unsigned char,3> ImageType;
 ImageType::Pointer im;
-int width, height, depth, imsize;
+long long width, height, depth, imsize;
 unsigned char *p;
 //unsigned short *label;
 unsigned int *label;
